Fix Background object, and place most WorldObject subclasses in one module.
