+++ pygfx/objects/__init__.py
@@ -1,9 +1,5 @@
 # flake8: noqa
 
 from ._base import WorldObject
-from ._group import Group
-from ._scene import Scene, Background
-from ._points import Points
-from ._line import Line
-from ._mesh import Mesh, InstancedMesh
-from ._volume import Volume
+from ._more import Group, Scene, Background, Points, Line, Mesh, Volume
+from ._instanced import InstancedMesh
