--- pygfx/objects/_group.py
@@ -1,7 +1,0 @@
-from ._base import WorldObject
-
-
-class Group(WorldObject):
-    """A Group is almost identical to a WorldObject. Its purpose is
-    to make working with groups of objects syntactically clearer.
-    """
