+++ pygfx/objects/_instanced.py
@@ -1,22 +1,12 @@
 import numpy as np
 
-from ._base import WorldObject
+from . import Mesh
 from ..resources import Buffer
 
 
-class Mesh(WorldObject):
-    """An object consisting of triangular faces, represented by vertices
-    (3D positions) and an index that defines the connectivity.
-
-    The picking info of a Mesh (the result of
-    ``renderer.get_pick_info()``) will for most materials include
-    ``instance_index`` (int), ``face_index`` (int), and ``face_coords``
-    (tuple of 3 floats). The latter are the barycentric coordinates for
-    each vertex of the face (with values 0..1).
-    """
-
-
 class InstancedMesh(Mesh):
+    """An instanced mesh with a matrix for each instance."""
+
     def __init__(self, geometry, material, count):
         super().__init__(geometry, material)
         count = int(count)
@@ -27,5 +17,6 @@ class InstancedMesh(Mesh):
         self.matrices = Buffer(matrices, nitems=count)
 
     def set_matrix_at(self, index: int, matrix):
+        """set the matrix for the instance at the given index."""
         matrix = np.array(matrix).reshape(4, 4)
         self.matrices.data[index] = matrix
