--- pygfx/objects/_points.py
@@ -1,11 +1,0 @@
-from ._base import WorldObject
-
-
-class Points(WorldObject):
-    """An object consisting of points represented by vertices (3D positions).
-
-    The picking info of a Points object (the result of
-    ``renderer.get_pick_info()``) will for most materials include
-    ``vertex_index`` (int).
-
-    """
